+++ datasentry-management/src/main/java/com/touhouqing/datasentry/cleaning/model/CleaningJob.java
@@ -22,7 +22,7 @@ public class CleaningJob {
 
 	private Long agentId;
 
-	private Long datasourceId;
+	private Integer datasourceId;
 
 	private String tableName;
